CBIDXT-100: CloseStream implemented.
- fix example app for dataport.
- changed application `opaque` type to uint16.
- feed's opaque value will increment dozen times a second, thus keeping it
  unique for a feed call and will not overflow for a long time.
- tools/multibuckets/multibuckets.go is passing.

Reviewed-on: http://review.couchbase.org/42105
Tested-by: Pratap Chakravarthy <prataprc@couchbase.com>

+++ secondary/projector/fake_upr.go
@@ -81,14 +81,13 @@ func (b *FakeBucket) GetChannel() <-chan *mc.UprEvent {
 
 // StartVbStreams is method receiver for BucketFeeder interface
 func (b *FakeBucket) StartVbStreams(
-	opaque uint32, ts *protobuf.TsVbuuid) (err error) {
-
+	opaque uint16, ts *protobuf.TsVbuuid) (err error) {
 	return
 }
 
 // EndVbStreams is method receiver for BucketFeeder interface
 func (b *FakeBucket) EndVbStreams(
-	opaque uint32, ts *protobuf.TsVbuuid) (err error) {
+	opaque uint16, ts *protobuf.TsVbuuid) (err error) {
 
 	return
 }
